requirements.txt
absl-py==2.1.0
aiohttp==3.10.9
alabaster==1.0.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
Babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.6
django-cors-headers==3.13.0
django-formtools==2.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==22.0.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
Pillow==10.3.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyMySQL==1.1.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
